+++ adapters/RPGRegionsAdapter/src/main/kotlin/me/ahdg6/typewriter/rpgregions/entries/event/DiscoverRegionEventEntry.kt
@@ -1,4 +1,4 @@
-package com.caleb.typewriter.rpgregions.entries.event
+package me.ahdg6.typewriter.rpgregions.entries.event
 
 import net.islandearth.rpgregions.api.events.RegionDiscoverEvent
 import lirand.api.extensions.server.server
