+++ documentation/docs/adapters/MythicMobsAdapter/MythicMobsAdapter.md
@@ -6,19 +6,19 @@ The MythicMobs Adapter is an adapter for the MythicMobs plugin. It allows you ha
 
 ### Action
 
-| Name                                                                    | Description      |
-|-------------------------------------------------------------------------|------------------|
-| [Spawn Mob Action](MythicMobsAdapter/entries/action/spawn_mob)   | Spawn a mob  |
-| [Despawn Mob Action](MythicMobsAdapter/entries/action/despawn_mob)             | Remove a mob       |
+| Name                                                                        | Description  |
+|-----------------------------------------------------------------------------|--------------|
+| [Spawn Mob Action](MythicMobsAdapter/entries/action/despawn_mythicmobs_mob) | Spawn a mob  |
+| [Despawn Mob Action](MythicMobsAdapter/entries/action/spawn_mythicmobs_mob) | Remove a mob |
 
 ### Event
 
-| Name                                                                    | Description      |
-|-------------------------------------------------------------------------|------------------|
-| [Mob Death Event](MythicMobsAdapter/entries/action/mob_death)   | When player kill a mob |
+| Name                                                                  | Description            |
+|-----------------------------------------------------------------------|------------------------|
+| [Mob Death Event](MythicMobsAdapter/entries/action/on_mythic_mob_die) | When player kill a mob |
 
 ### Facts
 
-| Name                                                         | Description                       |
-|--------------------------------------------------------------|-----------------------------------|
-| [Mob Exist Fact](MythicMobsAdapter/entries/fact/mob_exist_fact)       | Check if a mob exist |
+| Name                                                                   | Description          |
+|------------------------------------------------------------------------|----------------------|
+| [Mob Exist Fact](MythicMobsAdapter/entries/fact/mythic_mob_count_fact) | Check if a mob exist |
