+++ documentation/docs/adapters/MythicMobsAdapter/entries/action/despawn_mythicmobs_mob.mdx
