+++ documentation/docs/adapters/MythicMobsAdapter/entries/action/spawn_mythicmobs_mob.mdx
