+++ documentation/docs/adapters/MythicMobsAdapter/entries/event/on_mythic_mob_die.mdx
