+++ documentation/docs/adapters/MythicMobsAdapter/entries/fact/mythic_mob_count_fact.mdx
@@ -1,4 +1,4 @@
-import {FactsField, ReadonlyFactInfo} from "@site/src/components/EntryField";
+import {EntryField, FactsField, ReadonlyFactInfo} from "@site/src/components/EntryField";
 
 # Mob Count Fact
