+++ documentation/docs/adapters/RPGRegionsAdapter/RPGRegionsAdapter.md
@@ -6,19 +6,19 @@ The RPGRegions Adapter is an adapter for the RPGRegions plugin. It allows you to
 
 ### Action
 
-| Name                                                                    | Description      |
-|-------------------------------------------------------------------------|------------------|
-| [Discover Region Action](RPGRegionsAdapter/entries/action/discover_region)   | Discover a region  |
+| Name                                                                           | Description       |
+|--------------------------------------------------------------------------------|-------------------|
+| [Discover Region Action](RPGRegionsAdapter/entries/action/discover_rpg_region) | Discover a region |
 
 ### Events
 
-| Name                                                                  | Description                              |
-|-----------------------------------------------------------------------|------------------------------------------|
-| [Enter Region Event](RPGRegionsAdapter/entries/event/on_enter_region) | When a player enters a RPGRegions region |
-| [Exit Region Event](RPGRegionsAdapter/entries/event/on_exit_region)   | When a player exits a RPGRegions region  |
+| Name                                                                         | Description                              |
+|------------------------------------------------------------------------------|------------------------------------------|
+| [Enter Region Event](RPGRegionsAdapter/entries/event/on_discover_rpg_region) | When a player enters a RPGRegions region |
+| [Exit Region Event](RPGRegionsAdapter/entries/event/on_enter_rpg_region)     | When a player exits a RPGRegions region  |
 
 ### Facts
 
-| Name                                                         | Description                       |
-|--------------------------------------------------------------|-----------------------------------|
-| [In Region Fact](RPGRegionsAdapter/entries/fact/in_region_fact) | If the player is in a region |
+| Name                                                                | Description                  |
+|---------------------------------------------------------------------|------------------------------|
+| [In Region Fact](RPGRegionsAdapter/entries/fact/in_rpg_region_fact) | If the player is in a region |
