+++ documentation/docs/adapters/RPGRegionsAdapter/entries/action/discover_rpg_region.mdx
