+++ documentation/docs/adapters/RPGRegionsAdapter/entries/event/on_discover_rpg_region.mdx
