+++ documentation/docs/adapters/RPGRegionsAdapter/entries/fact/in_rpg_region_fact.mdx
